+++ src/actions/serve.js
@@ -21,7 +21,7 @@ const execa = require('execa')
 const path = require('path')
 const chalk = require('chalk')
 const os = require('os')
-const child_process = require('child_process')
+const { exec, execSync } = require('child_process')
 const isLocallyInstalled = require('../helpers/localinstallationcheck')
 const buildHelpers = require('../helpers/build')
 const sequence = require('../helpers/sequence')
@@ -31,18 +31,18 @@ module.exports = () => {
     () => buildHelpers.ensureLightningApp(),
     () => {
       const args = [
-        './build',
+        process.env.LNG_BUILD_FOLDER ? `./${process.env.LNG_BUILD_FOLDER}` : './build',
         process.env.LNG_SERVE_OPEN === 'false' ? false : '-o',
         process.env.LNG_SERVE_CACHE_TIME ? '-c' + process.env.LNG_SERVE_CACHE_TIME : '-c-1',
         process.env.LNG_SERVE_PORT ? '-p' + process.env.LNG_SERVE_PORT : false,
         process.env.LNG_SERVE_PROXY ? '-P' + process.env.LNG_SERVE_PROXY : false,
       ].filter(val => val)
 
-      const levelsDown = isLocallyInstalled() ? '../../../../..' : '../..'
-      const subprocess = execa(
-        path.join(__dirname, levelsDown, 'node_modules/.bin/http-server'),
-        args
-      )
+      const levelsDown = isLocallyInstalled()
+        ? buildHelpers.findFile(process.cwd(), 'node_modules/.bin/http-server')
+        : path.join(__dirname, '../..', 'node_modules/.bin/http-server')
+
+      const subprocess = execa(levelsDown, args)
 
       subprocess.catch(e => console.log(chalk.red(e.stderr)))
       subprocess.stdout.pipe(process.stdout)
@@ -50,9 +50,14 @@ module.exports = () => {
       // Hack for windows to prevent leaving orphan processes, resulting in multiple http-server running instances
       if (os.platform() === 'win32') {
         process.on('SIGINT', () => {
-          child_process.exec('taskkill /pid ' + subprocess.pid + ' /t /f', () => {
-            process.exit()
-          })
+          const task = 'taskkill /pid ' + subprocess.pid + ' /t /f'
+          if (process.env.LNG_LIVE_RELOAD) {
+            execSync(task)
+          } else {
+            exec(task, () => {
+              process.exit()
+            })
+          }
         })
       }